+++ src/main/java/com/kt/service/order/OrderService.java
@@ -40,7 +40,7 @@ public class OrderService {
 	private final DeliveryService deliveryService;
 	private final CartRepository cartRepository;
 	private final CartProductRepository cartProductRepository;
-	private final InventoryRepository inventoryRepository;
+	private final OrderValidator orderValidator;
 
 	// 주문번호 생성
 	private String generateOrderNumber() {
@@ -69,9 +69,8 @@ public class OrderService {
 		// 장바구니에 담긴 상품들 조회
 		List<CartProduct> cartProducts = cartProductRepository.findAllByCartId(cart.getId());
 
-		if (cartProducts.isEmpty()) {
-			throw new CustomException(ErrorCode.CART_EMPTY);
-		}
+		// 주문 검증 (장바구니 비어있는지, 재고 검증)
+		orderValidator.validateCartOrder(cartProducts);
 
 		Receiver receiver = new Receiver(
 			request.receiverName(),
@@ -84,15 +83,11 @@ public class OrderService {
 		Order order = Order.create(userId, receiver, orderNumber);
 		orderRepository.save(order);
 
-		// 재고 검증 및 주문 상품 생성
+		// 주문 상품 생성
 		for (CartProduct cartProduct : cartProducts) {
 			Product product = cartProduct.getProduct();
 			int quantity = cartProduct.getCount();
 
-			// 재고 검증
-			validateStock(product.getId(), quantity);
-
-			// 주문 상품 생성
 			OrderProduct orderProduct = OrderProduct.create(product, quantity, order);
 			orderProductRepository.save(orderProduct);
 			order.mapToOrder(orderProduct);
@@ -100,7 +95,7 @@ public class OrderService {
 		//총 금액 계산
 		order.calculateTotalAmount();
 
-		//TODO: 주문 완료 시점이 아닌 '결제 완료' 시점에 장바구니를 비우도록 변경하기
+		//TODO: 주문생성 시점이 아닌 '주문완료(결제 완료)' 시점에 장바구니를 비우도록 변경하기
 		cartProductRepository.deleteAllByCartId(cart.getId());
 
 		// 배송 정보 생성
@@ -109,14 +104,14 @@ public class OrderService {
 		return order;
 	}
 
-	// 2. 바로 주문
+	// 2. 바로 주문 생성
 	public Order createDirectOrder(Long userId, OrderRequest.DirectOrder request) {
 		// 상품 조회
 		Product product = productRepository.findById(request.productId())
 			.orElseThrow(() -> new CustomException(ErrorCode.PRODUCT_NOT_FOUND));
 
-		// 재고 검증
-		validateStock(product.getId(), request.quantity());
+		// 주문 검증 (재고 검증)
+		orderValidator.validateDirectOrder(product.getId(), request.quantity());
 
 		Receiver receiver = new Receiver(
 			request.receiverName(),
@@ -143,38 +138,26 @@ public class OrderService {
 		return order;
 	}
 
-	// 재고 검증
-	private void validateStock(Long productId, int requestedQuantity) {
-		Inventory inventory = inventoryRepository.findByProductId(productId)
-			.orElseThrow(() -> new CustomException(ErrorCode.INVENTORY_NOT_FOUND));
-
-		// 재고가 1 미만이면 주문 불가
-		if (inventory.getAvailable() < 1) {
-			throw new CustomException(ErrorCode.PRODUCT_STOCK_NOT_ENOUGH);
-		}
-
-		//TODO : 가용재고랑 실재고 중 어느걸 바탕으로 검증해야 하는지 고민해보기
-
-		// 요청 수량이 가용 재고보다 많으면 주문 불가
-		if (requestedQuantity > inventory.getAvailable()) {
-			throw new CustomException(ErrorCode.PRODUCT_STOCK_NOT_ENOUGH);
-		}
-	}
-
+	//내 주문 목록 조회
 	public List<Order> myOrderList(Long userId) {
 		return orderRepository.findByUserId(userId);
 	}
 
+	//내 주문 상세 조회
 	public Order myOrderInfo(String orderNumber, Long userId) {
 		return orderRepository.findByOrderNumberAndUserId(orderNumber, userId)
 			.orElseThrow(() -> new CustomException(ErrorCode.ORDER_NOT_FOUND));
 	}
 
+	//내 주문 취소('주문완료(결제완료)' 시점에서의 취소)
 	public Order cancelOrder(Long userId, String orderNumber) {
 		Order order = orderRepository.findByOrderNumberAndUserId(orderNumber, userId)
 			.orElseThrow(() -> new CustomException(ErrorCode.ORDER_NOT_FOUND));
 
 		order.cancel();
+
+		//주문 취소시 환불 요청
+		//재고 복구
 
 		return order;
 	}
